SPARK-48812 Remove group write permission from files included in .deb (#15)
Previously with a deb built from this repo, some files had the group write permission flag set which caused the expressway to complain:
2018-12-11T09:28:02.266+00:00 supervisor: UTCTime="2018-12-11 09:28:02,266" Module="developer.supervisor" Level="ERROR" CodeLocation="firestarter(197)" Error="Unexpected output" Process="/sbin/logrotated" Detail="error: Ignoring hybridserviceslog-rotate.conf because of bad file mode - must be 0644 or 0444. "

Although this no longer seemed to be the case when I built the deb, no intentional change was made to fix this so we're not fully sure what happened. To prevent this occurring again, this commit adds a build step that removes the group read permission for all files that will be included in the deb. It then asserts no files were found with that permission, and fails the build if this is not the case.

Also removing the old swpkg.lst file as it's no longer needed.

Jira link: https://jira-eng-gpk2.cisco.com/jira/browse/SPARK-48812