SPARK-52444 Add wait_for_install option for build & install script (#16)
This adds an option to the build_and_upgrade.sh script that waits until the newly built c_mgmt deb is successfully installed before exiting.

Passing the -w arg causes the script to poll the target expressway every second for 30 seconds to check the status of the installation. If it reports the new version was successfully installed within that time frame, the script finishes gracefully. If after 30 seconds the new deb hasn't been installed, the script logs the error as well as the current status of the c_mgmt deb on the target, and exits with a non-zero exit code. From local testing, the install consistently took about 12 seconds so 30 seconds should be plenty of leeway.

Jira link: https://jira-eng-gpk2.cisco.com/jira/browse/SPARK-52444