SPARK-29420 - Hybrid Management Federation Phase 2 Support (#132)
* Refactor to make it possible to inject dependencies

Refactor the u2c type to be able to inject the HTTP and DatabaseHandler dependencies. Rewrote the existing unit test by using Mock.

* SPARK-70448 - Consume new federation aware U2C APIs

Updating the template to point to the new Federation, phase 2 U2C user service URL, see https://wiki.cisco.com/display/WX2/U2C+Federation+Phase-2 for details.

Note: there is a difference in implementation, as we now pass in the services in the service map as a URL param: This way we don't get the URL pointers for ALL services, but for the services we do care about (those in our service map).

Note that I have kept the test where we update the CDB with only the values in our service map. The reason to keep this is defensive: If U2C for a reason stop to respect the URL param and return the catalog for all service, we still filter in our code ...

* SPARK-68250 - Always set the U2C URL in CDB

Now make sure we do check if the U2C host url is set in the DB (not just in the configuration/template, as this has a default value ...).

This will ensure consumers of data directly from CDB will get at least the default value we have in the template.

When this fix gets deployed, we will do the check every time the U2C type gets initiated: in the `u2cthread` run method; then we will set a flag, preventing us to do it again. A very defensive approach would be to remove that condition (always do the check, every 24 hours), AND also do a read after write to see that the value is TRULY set, but I will leave these decisions to the people who know this code :)

* Fixing logging statement

* Refactoring