Port test_bootstrap_fuse_flow (#146)
This ports test_bootstrap_fuse_flow, which UI tests the bootstrapping flow. Note that this only covers the bootstrapping part: I've intentionally left out the rest of the fuse flow, as it is already covered by the integration UI tests.

Also note that this adds a new Jenkins step after deploying to latest. This is done because the prevent upgrade flag interacts with the bootstrap UI flow, making us unable to test the bootstrap flow if the flag is set.